+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Version 0.8.0-SNAPSHOT - Build 20161220
 - Removing the need to call KnowledgeBase.init() in any predict/transform method.
 - Improved DatabaseConnector: existsObject method, InMemory now stores objects independently, MapDB stores all files in folder.
 - The training parameters are now provided on the constructor of the algorithms not with a setter.
+- TextClassifier inherits from Modeler.
+- Removed all unnecessary passing of class objects from Stepwise Regression, Wrappers and Ensumble learning classes.
+- Removed the ignoringNumericalFeatures variable from AbstractCategoricalFeatureSelector.
 
 Version 0.7.1-SNAPSHOT - Build 20161217
 ---------------------------------------